updated AdminServices layer

+++ src/main/java/com/nk/service/AdminServiceImpl.java
@@ -6,20 +6,26 @@ import com.nk.dto.DepartmentRequestDto;
 import com.nk.dto.DepartmentResponseDto;
 import com.nk.exception.DepartmentNotFoundException;
 import com.nk.exception.EmployeeNotFoundException;
+import com.nk.exception.InvalidOperation;
 import com.nk.mapper.DepartmentMapper;
 import com.nk.repository.DepartmentRepository;
 import com.nk.repository.EmployeeRepository;
+import com.nk.validators.DepartmentValidators;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.time.LocalDate;
 
 
 @Service
 @RequiredArgsConstructor
 public class AdminServiceImpl implements AdminServices{
 
+
     private final DepartmentRepository repo;
     private final EmployeeRepository employeeRepository;
+    private final DepartmentValidators validators;
 
 
     @Transactional
@@ -28,6 +34,8 @@ public class AdminServiceImpl implements AdminServices{
         if (!repo.existsByDeptName(dto.getDeptName())) {
             Department department = DepartmentMapper.toDepartment(dto);
             department.setDeptCode(generateDeptCode());
+            department.setCreatedAt(LocalDate.now());
+            department.setUpdatedAt(LocalDate.now());
             repo.save(department);
             return  DepartmentMapper.toDepartmentResponseDto(department);
         }else {
@@ -37,11 +45,15 @@ public class AdminServiceImpl implements AdminServices{
 
     @Override
     public DepartmentResponseDto getDepartmentById(Long id) {
-        return null;
+        Department department = repo.findById(id).orElseThrow(() -> new DepartmentNotFoundException("Department not found with id " + id));
+        DepartmentResponseDto responseDto = DepartmentMapper.toDepartmentResponseDto(department);
+        return responseDto;
     }
 
     @Override
     public DepartmentResponseDto getDepartmentByDeptCode(String deptCode) {
+
+
         Department dept = repo.getDepartmentByDeptCode(deptCode).
                 orElseThrow(() -> new DepartmentNotFoundException("Department not found with : "+deptCode));
 
@@ -50,16 +62,34 @@ public class AdminServiceImpl implements AdminServices{
 
     @Transactional
     @Override
-    public DepartmentResponseDto updateDepartment(DepartmentResponseDto department) {
-        return null;
+    public DepartmentResponseDto updateDepartment(String deptCode,DepartmentRequestDto requestDto) {
+        Department department=repo.getDepartmentByDeptCode(deptCode).orElseThrow(() -> new DepartmentNotFoundException("Department not found with id "+deptCode));
+
+        validators.validateDepartmentRequestDto(requestDto); //validate request
+
+        department.setDeptName(requestDto.getDeptName()); // assign new values
+        department.setLocation(requestDto.getLocation());
+        department.setManagerName(requestDto.getManagerName());
+        department.setUpdatedAt(LocalDate.now()); // dirty checking
+
+
+        DepartmentResponseDto response = DepartmentMapper.toDepartmentResponseDto(department);
+
+        return response;
     }
 
     @Transactional
     @Override
-    public String deleteDepartment(String deptCode) {
+    public String deleteDepartment(String deptCode)throws Exception {
       Department dept=  repo.getDepartmentByDeptCode(deptCode)
               .orElseThrow(() -> new DepartmentNotFoundException("Department not found with : "+deptCode));
-      repo.delete(dept);
+
+        Long count = employeeRepository.countEmployeesByDepartment_Id(dept.getDepartmentId());
+        if (count>0){
+            throw new InvalidOperation("still "+count+" employees are working in "+deptCode+" department so you can't delete it");
+        }
+
+        repo.delete(dept);
       return dept.getDeptName();
     }
 
@@ -76,6 +106,7 @@ public class AdminServiceImpl implements AdminServices{
         return "successfully hiked the salary for "+empCode;
     }
 
+    @Transactional
     @Override
     public String transferDepartment(String empCode,String deptCode){
         Department dept=repo.getDepartmentByDeptCode(deptCode)
@@ -83,7 +114,7 @@ public class AdminServiceImpl implements AdminServices{
         Employee emp=employeeRepository.getEmployeeByEmpCode(empCode)
                 .orElseThrow(() -> new EmployeeNotFoundException("Employee not found with : "+empCode));
 
-        emp.getDepartment().setDepartmentId(dept.getDepartmentId()); // Hibernate Dirty read
+        emp.setDepartment(dept); // Hibernate checking
 
         //employeeRepository.transferEmployeeToDepartment(emp.getEmpCode(),dept.getDepartmentId());
         return "successfully transferred "+empCode+" to "+dept.getDeptName();
